run.command(): Fix operation of -continue
The existing implementation of -continue assumed that, for a list of elements to be provided to subprocess.run(), each element was a string, and therefore string operations could be applied. However with the adoption of pathlib for handling user-specified paths in #2678, an element in this list may be a class derived from those provided in the pathlib module that does not provide string functions.