mrfilter: New operation demodulate
Performs demodulation of a complex image. Low-frequency phase map is estimated by performing an FFT back to k-space and manipulating the resulting data. By default, a Hann window is applied to the k-space data, to which an iFFT is applied to provide an estimate of the low-frequency phase variations. Alternatively, strictly linear phase ramps can be estimated, by finding the location and phase of the (interpolated) k-space maximum.