Add lightweight SpinLock class
This provides a more lightweight alternative to a full-blown mutex,
useful in cases where there may be a very large number of small objects
that each need their own thread protection. It relies on
std::atomic_flag, and only takes up a single byte. It should only be
used sparingly and in cases where collisions are expected to be rare,
since spinning threads will keep running.