tcksift2: Fix to differential mode fixel-wise regularisation
If all streamlines ascribed to a fixel / streamline group attain a delta coefficient of +-1.0, and that becomes the reference value for that fixel / streamline group, then the cost function for regularising by delta coefficient with respect to that reference experiences a discontinuity. This change should yield sensible values when that occurs.